backend/requirements.txt
# Business Guardian AI - Python Dependencies

# =====================================
# CONFLUENT / KAFKA
# =====================================
confluent-kafka==2.3.0
fastavro==1.9.0

# =====================================
# GOOGLE CLOUD
# =====================================
google-cloud-aiplatform==1.38.0
google-cloud-bigquery==3.14.0
google-cloud-storage==2.13.0
google-cloud-firestore==2.13.1
google-generativeai==0.3.1
firebase-admin==6.3.0

# =====================================
# WEB FRAMEWORK
# =====================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# =====================================
# DATA PROCESSING
# =====================================
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2

# =====================================
# AUTHENTICATION & SECURITY
# =====================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
PyJWT==2.8.0
cryptography==41.0.7

# =====================================
# PAYMENT PROCESSING
# =====================================
stripe==7.8.0

# =====================================
# HTTP & API CLIENTS
# =====================================
httpx==0.25.1
requests==2.31.0
aiohttp==3.9.1

# =====================================
# CACHING & QUEUING
# =====================================
redis==5.0.1
celery==5.3.4

# =====================================
# MONITORING & LOGGING
# =====================================
python-json-logger==2.0.7
structlog==23.2.0
sentry-sdk==1.38.0

# =====================================
# TESTING
# =====================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.1
faker==20.1.0

# =====================================
# CODE QUALITY
# =====================================
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

# =====================================
# UTILITIES
# =====================================
python-dateutil==2.8.2
pytz==2023.3
uuid==1.30
